typescript(vx-geo): export useful consumer types from new types index
