+++ packages/vx-geo/src/projections/NaturalEarth.tsx
@@ -1,5 +1,6 @@
 import React from 'react';
-import Projection, { ProjectionProps, GeoPermissibleObjects } from './Projection';
+import Projection, { ProjectionProps } from './Projection';
+import { GeoPermissibleObjects } from '../types';
 
 /**
  * All props pass through to `<Projection projection="naturalEarth" {...props} />`
