+++ packages/vx-geo/src/types.ts
@@ -1,0 +1,14 @@
+import { GeoProjection, GeoPermissibleObjects as GeoPermissibleObjectType } from 'd3-geo';
+
+export type GeoPermissibleObjects = GeoPermissibleObjectType;
+
+export type Projection = ProjectionPreset | (() => GeoProjection);
+
+// @TODO: Implement all projections of d3-geo
+export type ProjectionPreset =
+  | 'orthographic'
+  | 'albers'
+  | 'albersUsa'
+  | 'mercator'
+  | 'naturalEarth'
+  | 'equalEarth';
